Fix documentation errors and add supported data types
Eventually we will want to list the supported data types somewhere in
the UI. For now I've created a constant which lists every type the
exporter can work with. I also use it as the default when no data_types
are provided.

I also updated some out of date documentation :anguished: